update the item entities

+++ src/main/java/saveDataEditor/ItemEntities/ItemEntity.java
@@ -3,14 +3,14 @@ package saveDataEditor.ItemEntities;
 public interface ItemEntity {
      long id=0;
      String name="";
-     Long amount=1L;
+     Double amount=1d;
      Long price=0L;
      Long age = 0L;
      Double quality = 0d;
 
       long getId();
       String getName();
-      Long getAmount();
+      Double getAmount();
       Long getPrice();
       Long getAge();
       Double getQuality();
