+++ src/main/java/saveDataEditor/ItemEntities/ResourceInformation.java
@@ -3,12 +3,12 @@ package saveDataEditor.ItemEntities;
 public class ResourceInformation implements ItemEntity{
     private long id;
     private String name;
-    private Long amount;
+    private Double amount;
     private Long price;
     private Long age;
     private Double quality;
 
-    public ResourceInformation(long id, String name, Long amount, Long price){
+    public ResourceInformation(long id, String name, Double amount, Long price){
         this.id=id;
         this.name=name;
         this.amount=amount;
@@ -23,11 +23,11 @@ public class ResourceInformation implements ItemEntity{
         return this.name == null ? "not a resource" : name;
     }
 
-    public Long getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
